@@ -1,26 +1,26 @@
 name: flutter_svg
 description: An SVG rendering and widget library for Flutter, which allows painting and displaying Scalable Vector Graphics 1.1 files.
 homepage: https://github.com/dnfield/flutter_svg
-version: 0.18.1+1
+version: 0.19.1+1
 
 dependencies:
-  path_drawing: ^0.4.1
-  xml: ^4.2.0
-  vector_math: ^2.0.8
-  meta: ^1.1.8
   flutter:
     sdk: flutter
+  meta: ^1.3.0-nullsafety.3
+  path_drawing: ^0.4.1+1
+  vector_math: ^2.1.0-nullsafety.3
+  xml: ^4.5.1
 
 dev_dependencies:
+  flutter_test:
+    sdk: flutter
   mockito: ^4.1.1
   path: ^1.7.0
   test: ^1.14.3
-  flutter_test:
-    sdk: flutter
 
 environment:
   sdk: ">=2.2.0 <3.0.0"
-  flutter: ">=1.18.0-6.0.pre <2.0.0"
+  flutter: ">=1.22.0 <2.0.0"
 
 ## For developing features that require new functionality in engine:
 # dependency_overrides:
