@@ -1,4 +1,3 @@
-import 'dart:async';
 import 'dart:collection';
 import 'dart:ui';
 
@@ -66,6 +65,7 @@ class _TextInfo {
     this.offset,
     this.transform,
   );
+
   final DrawableStyle style;
   final Offset offset;
   final Matrix4 transform;
@@ -77,6 +77,8 @@ class _TextInfo {
 class _Elements {
   static Future<void> svg(SvgParserState parserState) {
     final DrawableViewport viewBox = parseViewBox(parserState.attributes);
+    final String id = parserState.attribute('id', def: '');
+
     // TODO(dnfield): Support nested SVG elements. https://github.com/dnfield/flutter_svg/issues/132
     if (parserState._root != null) {
       FlutterError.reportError(FlutterErrorDetails(
@@ -96,6 +98,7 @@ class _Elements {
         _SvgGroupTuple(
           'svg',
           DrawableGroup(
+            id,
             <Drawable>[],
             parseStyle(
               parserState.attributes,
@@ -109,6 +112,7 @@ class _Elements {
       return null;
     }
     parserState._root = DrawableRoot(
+      id,
       viewBox,
       <Drawable>[],
       parserState._definitions,
@@ -126,6 +130,7 @@ class _Elements {
   static Future<void> g(SvgParserState parserState) {
     final DrawableParent parent = parserState.currentGroup;
     final DrawableGroup group = DrawableGroup(
+      parserState.attribute('id', def: ''),
       <Drawable>[],
       parseStyle(
         parserState.attributes,
@@ -145,6 +150,7 @@ class _Elements {
   static Future<void> symbol(SvgParserState parserState) {
     final DrawableParent parent = parserState.currentGroup;
     final DrawableGroup group = DrawableGroup(
+      parserState.attribute('id', def: ''),
       <Drawable>[],
       parseStyle(
         parserState.attributes,
@@ -183,6 +189,7 @@ class _Elements {
     final DrawableStyleable ref =
         parserState._definitions.getDrawable('url($xlinkHref)');
     final DrawableGroup group = DrawableGroup(
+      parserState.attribute('id', def: ''),
       <Drawable>[ref?.mergeStyle(style)],
       style,
       transform: transform.storage,
@@ -486,6 +493,7 @@ class _Elements {
     final DrawableParent parent = parserState._parentDrawables.last.drawable;
     final DrawableStyle parentStyle = parent.style;
     final DrawableRasterImage drawable = DrawableRasterImage(
+      parserState.attribute('id', def: ''),
       image,
       offset,
       parseStyle(
@@ -537,6 +545,7 @@ class _Elements {
       );
       parserState.currentGroup.children.add(
         DrawableText(
+          parserState.attribute('id', def: ''),
           fill,
           stroke,
           lastTextInfo.offset,
@@ -838,6 +847,7 @@ class SvgParserState {
     final DrawableStyle parentStyle = parent.style;
     final Path path = pathFunc(attributes);
     final DrawableStyleable drawable = DrawableShape(
+      getAttribute(attributes, 'id', def: ''),
       path,
       parseStyle(
         attributes,
